+++ Sources/DeckKit/Favorites/UserDefaultsFavoriteService.swift
@@ -14,39 +14,28 @@ import Foundation
  */
 public class UserDefaultsFavoriteService: FavoriteService {
     
-    /**
-     Create a service instance.
-     
-     - Parameters:
-       - defaults: The store to persist data in.
-     */
-    public init(defaults: UserDefaults = .standard) {
+    /// Create a service instance.
+    public init(
+        defaults: UserDefaults = .standard
+    ) {
         self.defaults = defaults
     }
     
-    
     private let defaults: UserDefaults
+}
+
+public extension UserDefaultsFavoriteService {
     
-    
-    /**
-     Get all favorite item IDs.
-     */
-    public func getFavorites<Item: Favoritable>(for type: Item.Type) -> [Item.ID] {
+    func getFavorites<Item: Identifiable>(for type: Item.Type) -> [Item.ID] {
         let key = self.key(for: type)
         return defaults.array(forKey: key) as? [Item.ID] ?? []
     }
     
-    /**
-     Check whether or not a certain item is a favorite.
-     */
-    public func isFavorite<Item: Favoritable>(_ item: Item) -> Bool {
+    func isFavorite<Item: Identifiable>(_ item: Item) -> Bool {
         getFavorites(for: Item.self).contains(item.id)
     }
     
-    /**
-     Set whether or not a certain item is a favorite.
-     */
-    public func setIsFavorite<Item: Favoritable>(_ isFavorite: Bool, for item: Item) {
+    func setIsFavorite<Item: Identifiable>(_ isFavorite: Bool, for item: Item) {
         var favorites = getFavorites(for: Item.self)
         if isFavorite {
             favorites.append(item.id)
@@ -56,17 +45,14 @@ public class UserDefaultsFavoriteService: FavoriteService {
         defaults.set(favorites, forKey: key(for: Item.self))
     }
     
-    /**
-     Toggle whether or not a certain item is a favorite.
-     */
-    public func toggleIsFavorite<Item: Favoritable>(for item: Item) {
+    func toggleIsFavorite<Item: Identifiable>(for item: Item) {
         setIsFavorite(!isFavorite(item), for: item)
     }
 }
 
 private extension UserDefaultsFavoriteService {
     
-    func key<Item: Favoritable>(for type: Item.Type) -> String {
+    func key<Item: Identifiable>(for type: Item.Type) -> String {
         "com.danielsaidi.deckkit.favorites.\(String(describing: type))"
     }
 }